refactor: #19 AuthenticationToken 삭제
