+++ src/main/java/site/dogether/common/config/web/resolver/AuthenticationArgumentResolver.java
@@ -6,7 +6,6 @@ import org.springframework.web.bind.support.WebDataBinderFactory;
 import org.springframework.web.context.request.NativeWebRequest;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.method.support.ModelAndViewContainer;
-import site.dogether.auth.infrastructure.AuthenticationToken;
 
 @Component
 public class AuthenticationArgumentResolver implements HandlerMethodArgumentResolver {
